Adds test to conditionally initialize the repository (#53)
- Removes unnecessary provision and initialize input variables
- Adds check for config.yaml file in the gitops repo before initializing. If the file exists then initialization is skipped
- Bumps git-repo submodule to v1.6.2

Closes #53

+++ scripts/initialize-gitops.sh
@@ -25,6 +25,11 @@ git clone "https://${TOKEN}@${REPO}" .tmpgitops
 
 cd .tmpgitops || exit 1
 
+if [[ -f config.yaml ]]; then
+  echo "Repository already initialized. Nothing to do"
+  exit 0
+fi
+
 export BRANCH=$(git rev-parse --abbrev-ref HEAD)
 
 cp -R "${TEMPLATE_DIR}"/* .
